+++ package.json
@@ -98,7 +98,8 @@
     "disabled": false,
     "dependencies": {
         "@hackolade/fetch": "1.3.0",
-        "graphql": "16.10.0"
+        "graphql": "16.10.0",
+        "ip": "2.0.1"
     },
     "lint-staged": {
         "*.{js,ts,json}": "prettier --write"
